Refactor shared constants, utilities, and types into reusable modules
Extract duplicated code into shared locations:
- src/utils/constants.ts: SOURCE_PATTERNS, IGNORE_PATTERNS, TEST_FILE_RE
- src/utils/text.ts: countLines() utility
- src/types/report.ts: RepoProfile interface

Both fileDiscovery.ts and loc.ts now import from the shared modules
instead of defining their own copies.
