+++ src/collect/loc.ts
@@ -2,53 +2,18 @@
  * Repository profiling module.
  *
  * Computes high-level repository statistics — file counts by type and lines
- * of code (total, source, test) — before AST analysis runs. Reuses the same
- * ignore list as file discovery so counts stay consistent.
+ * of code (total, source, test) — before AST analysis runs. Uses shared
+ * constants so ignore patterns and file classification stay consistent
+ * across all modules.
  */
 
 import { readFile } from "node:fs/promises";
 import fg from "fast-glob";
+import { SOURCE_PATTERNS, IGNORE_PATTERNS, TEST_FILE_RE } from "../utils/constants.js";
+import { countLines } from "../utils/text.js";
+import type { RepoProfile } from "../types/report.js";
 
-const SOURCE_PATTERNS = ["**/*.ts", "**/*.tsx"];
-
-const IGNORE = [
-  "**/node_modules/**",
-  "**/dist/**",
-  "**/build/**",
-  "**/.next/**",
-  "**/out/**",
-  "**/coverage/**",
-  "**/.git/**",
-];
-
-const TEST_FILE_RE = /\.(test|spec)\.(ts|tsx)$/;
-
-/** Breakdown of file counts and lines of code for a repository. */
-export interface RepoProfile {
-  totalFiles: number;
-  tsFiles: number;
-  tsxFiles: number;
-  testFiles: number;
-  totalLOC: number;
-  sourceLOC: number;
-  testLOC: number;
-}
-
-/**
- * Count the number of newline-delimited lines in a string.
- *
- * @param text - Raw file contents.
- * @returns Line count (empty file = 0, single trailing newline not double-counted).
- */
-function countLines(text: string): number {
-  if (text.length === 0) return 0;
-  let count = 1;
-  for (let i = 0; i < text.length; i++) {
-    if (text[i] === "\n") count++;
-  }
-  if (text[text.length - 1] === "\n") count--;
-  return count;
-}
+export type { RepoProfile } from "../types/report.js";
 
 /**
  * Profile a repository's TypeScript / TSX source files.
@@ -63,7 +28,7 @@ export async function profileRepo(repoPath: string): Promise<RepoProfile> {
   const files = await fg(SOURCE_PATTERNS, {
     cwd: repoPath,
     absolute: true,
-    ignore: IGNORE,
+    ignore: IGNORE_PATTERNS,
   });
 
   let tsFiles = 0;
